+++ packages/playground-react/pages/avatar/avatar.tsx
@@ -2,6 +2,22 @@ import { Avatar, AvatarGroup, FlexBox, Text } from '@cos-ui/react';
 
 import AVATARS from '@constants/avatar';
 
+const dummyAvatarsDouble = new Array(10)
+  .fill(0)
+  .map(() => AVATARS)
+  .flat()
+  .map((avatar, index) => ({ ...avatar, id: index }));
+const dummyAvatarsTriple = new Array(100)
+  .fill(0)
+  .map(() => AVATARS)
+  .flat()
+  .map((avatar, index) => ({ ...avatar, id: index }));
+const dummyAvatarsFour = new Array(1000)
+  .fill(0)
+  .map(() => AVATARS)
+  .flat()
+  .map((avatar, index) => ({ ...avatar, id: index }));
+
 const AvatarPage = () => (
   <>
     <FlexBox sx={{ flexDirection: 'column', gap: 1, mb: 2 }}>
@@ -14,7 +30,7 @@ const AvatarPage = () => (
     </FlexBox>
     <FlexBox sx={{ flexDirection: 'column', gap: 1 }}>
       <Text variant="sectionTitle">Avatar Group</Text>
-      <AvatarGroup flexAlign="flex-end" max={3} spacing="small">
+      <AvatarGroup max={3} spacing="small">
         {AVATARS.slice(0, 2).map((avatar) => (
           <Avatar
             key={avatar.id}
@@ -24,8 +40,38 @@ const AvatarPage = () => (
           />
         ))}
       </AvatarGroup>
-      <AvatarGroup flexAlign="flex-end" max={3} spacing="small">
+      <AvatarGroup max={3} spacing="small">
         {AVATARS.map((avatar) => (
+          <Avatar
+            key={avatar.id}
+            src={avatar.src}
+            alt={avatar.name}
+            sx={{ width: '4.2rem', height: '4.2rem' }}
+          />
+        ))}
+      </AvatarGroup>
+      <AvatarGroup max={3} spacing="small">
+        {dummyAvatarsDouble.map((avatar) => (
+          <Avatar
+            key={avatar.id}
+            src={avatar.src}
+            alt={avatar.name}
+            sx={{ width: '4.2rem', height: '4.2rem' }}
+          />
+        ))}
+      </AvatarGroup>
+      <AvatarGroup max={3} spacing="small">
+        {dummyAvatarsTriple.map((avatar) => (
+          <Avatar
+            key={avatar.id}
+            src={avatar.src}
+            alt={avatar.name}
+            sx={{ width: '4.2rem', height: '4.2rem' }}
+          />
+        ))}
+      </AvatarGroup>
+      <AvatarGroup max={3} spacing="small">
+        {dummyAvatarsFour.map((avatar) => (
           <Avatar
             key={avatar.id}
             src={avatar.src}
